requirements.txt
torch>=1.9.0
numpy>=1.19.0
pandas>=1.1.0
pyyaml>=5.3
scikit-learn>=0.24.0
matplotlib>=3.3.0
